requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python_crontab==2.7.1
requests==2.28.1
rich==13.4.2
selenium==4.9.1
webdriver_manager==3.8.6
ytmusicapi==1.0.2
